src/nymbus/package/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
docker==6.0.1
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
packaging==23.0
PyYAML==6.0
requests==2.28.2
smmap==5.0.0
typing-extensions==4.4.0
urllib3==1.26.14
websocket-client==1.5.1
